Add a feature flag to control web app window experiment cleanup phase.
When the experiment ends, cleanup code runs that persists state to the
web app DB. Past experience tells us that running code when a flag is
disabled is dangerous, as there is no way to prevent that code running
if the flag needs to be disabled for some reason. So this adds a second
base::feature flag that, if disabled, prevents the cleanup/persistence
code from running.

The flag is enabled by default, but can be disabled by finch if needed.

Bug: 1385246
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4344393
Commit-Queue: Glen Robertson <glenrob@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1119158}

+++ chrome/browser/web_applications/preinstalled_web_app_window_experiment.cc
@@ -144,6 +144,10 @@ void PersistStateFromPrefsToWebAppDb(PrefService* pref_service,
 
 }  // namespace
 
+BASE_FEATURE(kWebAppWindowExperimentCleanup,
+             "WebAppWindowExperimentCleanup",
+             base::FEATURE_ENABLED_BY_DEFAULT);
+
 PreinstalledWebAppWindowExperiment::PreinstalledWebAppWindowExperiment(
     Profile* profile)
     : profile_(profile) {
@@ -265,10 +269,12 @@ void PreinstalledWebAppWindowExperiment::CleanUp() {
   registrar_observation_.Reset();
   preferred_apps_observation_.Reset();
 
-  PersistStateFromPrefsToWebAppDb(profile_->GetPrefs(),
-                                  *WebAppProvider::GetForWebApps(profile_));
+  if (base::FeatureList::IsEnabled(kWebAppWindowExperimentCleanup)) {
+    PersistStateFromPrefsToWebAppDb(profile_->GetPrefs(),
+                                    *WebAppProvider::GetForWebApps(profile_));
 
-  utils::DeleteExperimentPrefs(profile_->GetPrefs());
+    utils::DeleteExperimentPrefs(profile_->GetPrefs());
+  }
 
   setup_done_for_testing_.Signal();
 }
